Add entity and array updaters for normalized data persistence
Through testing we realized that normalizing the data persisted to
the database is not sufficient. It's possible that the endereco plugin
changes the in-memory entity, that will be used by the Shopware
system for persistence later. E.g. when customer address is
transformed into order address, apparently the entity is used. It's
transformed into array, then an order address is created and finally
it's persisted when the order is persisted.

Now because the entity might receive a too long text from endereco
plugin (e.g. setStreet would set a string longer than 255), the order
would not be able to be created, as it would run into database error.

Customer Address Entity → Endereco Plugin (street > 255 chars)
                     ↓
                 Order Address Creation
                     ↓
                 Array Transformation
                     ↓
                 Database Persistence (constraint violation)

Consequently we need to also normalize the data we write to in-memory
entities and by extension of this logic, to post data array, that we
also modify. For that in this commit we created updater service
classes, that take a normalized payload (that we used to update
database) and use it to update entity or array fields. This should
provide more stability downstream.

Ref: https://mobilemojo.atlassian.net/browse/DEV-60?focusedCommentId=38338